feat: Condense friends and search functionality onto one page
