+++ pages/friends.tsx
@@ -1,12 +1,25 @@
 import React from "react";
-import { Card, Col, Container, Row } from "react-bootstrap";
+import { Button, Card, Col, Container, Form, Row } from "react-bootstrap";
+import { useDebouncedCallback } from "use-debounce";
 import { useUser } from "../lib/hooks/auth";
-import { useFriends } from "../lib/hooks/profile";
+import { searchProfiles, setFriend, useFriends } from "../lib/hooks/profile";
 import IUser from "../types/IUser";
 
-export default function Search() {
+export default function Friends() {
   const { user } = useUser();
   const { friends } = useFriends();
+
+  const [queriedProfiles, setQueriedProfiles] = React.useState<IUser[]>([]);
+  const [searchQuery, setSearchQuery] = React.useState<string>("");
+
+  const debounced = useDebouncedCallback((value: string) => {
+    setSearchQuery(value);
+    if (value.length >= 2) {
+      searchProfiles(value).then((profiles: IUser[]) => {
+        setQueriedProfiles(profiles);
+      });
+    }
+  }, 500);
 
   if (!user) {
     return (
@@ -24,17 +37,26 @@ export default function Search() {
     );
   }
 
+  const listToMap = searchQuery ? queriedProfiles : friends;
+
   return (
     <Container className="text-center mt-5">
       <h1>Friends</h1>
       <br />
       <h4>See how your friends are doing on Wordify lately.</h4>
       <h5>
-        Add new friends by visiting the &quot;Search for Friends&quot; page.
+        Add new friends by searching for user profiles by their first and/or
+        last name.
       </h5>
       <br />
+      <Form.Control
+        type="text"
+        placeholder="Search"
+        onChange={(e) => debounced(e.target.value)}
+      />
+      <br />
       <Row xs={1} md={2} lg={3} xl={4}>
-        {friends.map((profile: IUser) => {
+        {listToMap.map((profile: IUser) => {
           let averageTime = 0;
           for (const time of profile.pastTimes) {
             if (time) {
@@ -74,6 +96,29 @@ export default function Search() {
             <>No attempt yet today</>
           );
 
+          const isSelf = profile._id === user._id;
+          const isFriend = (user.friends as string[]).includes(profile._id);
+
+          const FriendButton = isFriend ? (
+            <Button
+              variant="secondary"
+              size="sm"
+              onClick={() => {
+                setFriend(profile._id, false);
+              }}
+            >
+              Unfriend
+            </Button>
+          ) : (
+            <Button
+              variant="secondary"
+              size="sm"
+              onClick={() => setFriend(profile._id, true)}
+            >
+              Add Friend
+            </Button>
+          );
+
           return (
             <Col key={profile._id}>
               <Card>
@@ -92,6 +137,7 @@ export default function Search() {
                     <br />
                     {TodaysAttempt}
                   </Card.Text>
+                  {!isSelf && FriendButton}
                 </Card.Body>
               </Card>
             </Col>
